[experiment] Allocate all memory with cuMemCreate for zero-copy NCCL
Until NCCL can support expandable segments, we need a temporary way to
ensure all memory is allocated with cuMemCreate but not allocation spans
multiple cuMemCreate segments. This is a hack to do that which uses
the expandable segments pathway to do the memory allocation but
sets address_space_size == segment_size so the segments do not expand
and are a single cuMemCreate allocation.

ghstack-source-id: c0efc288a9f4aa454fe83e0aca730bcff765c40a
Pull Request resolved: https://github.com/pytorch/pytorch/pull/128573