[C10D] Separate deviceKey from streamKey in getNCCLComm
The paves the way for a set of changes that would let us use the same
nccl communicator for collectives and p2p ops (and have that be eagerly
initialized instead of lazily initialized), but always use a dedicated
nccl stream for p2p operations with different ranks to ensure they can
overlap.

This PR only adds a new streamKey that defaults to the value of the
deviceKey, so should not affect runtime behavior at all.

ghstack-source-id: a7cb2b99ad60e24db6a3ddb3610a144a98015191
Pull Request resolved: https://github.com/pytorch/pytorch/pull/129435