[HOP] improve naming for subgraph inputs
Previously, subgraph input names were whatever the input proxies were,
which were confusing. This PR changes those names to be
whatever the names of the arguments the functions being
speculate_subgraph'ed are. This is best-effort: if we can't figure it
out then we go back to the previous strategy.

Test Plan:
- existing expecttests

ghstack-source-id: 9a8ba5279d2572f98fff89d2990d62e32d07e83d
Pull Request resolved: https://github.com/pytorch/pytorch/pull/130255