[decomps] Fix aten._to_copy decomp
`aten._to_copy` can receive a python number as input. This occurs in
torch.compile support for vmap (see #130188). Previously, this would
raise an assertion error. This PR changes it so that if we see a python
number, we call torch.scalar_tensor on it first (h/t bdhirsh).

Fixes #130362

Fixes #130188

Test Plan:
- new tests

ghstack-source-id: d6c9b286a9bf7ceef7cc16b012a7f830c8cf8eac
Pull Request resolved: https://github.com/pytorch/pytorch/pull/130381