[pipelining] Refactor test_schedule to fix "-k"
This is kind of a short-sighted workaround and we should actually come
up with a way to fix this in general, but I got annoyed that I can't use
-k to filter tests in test_schedule, and realized it's because we jam
tests using the new MultiProcContinuousTest fixture together with
old-style tests.

For now I separate the two types of tests so -k works again.

ghstack-source-id: 47b9d54ce8b5f8db4c175115a3f2400ae34178cf
Pull Request resolved: https://github.com/pytorch/pytorch/pull/130294