[NJT] throw an exception if nested_tensor_from_jagged is fx-traced without being fx.wrapped
The NJT constructor can't be fx-traced safely due to the dummy nt used:

https://github.com/pytorch/pytorch/blob/774ca93fd2b99808761010528267ef14fa816bda/torch/nested/_internal/nested_tensor.py#L501-L508

The error doesn't appear immediately, but appears if you try to move a module with an fx-traced NJT constructor onto a different device, or try to serialize it. Let's throw an error if we try to fx-trace the NJT constructor so users know to wrap the call.

ghstack-source-id: 8dfa8706728e50ef367b0837913b0abb3a14fa53
Pull Request resolved: https://github.com/pytorch/pytorch/pull/130702