[custom_op] triton_op API V0
This is the initial version of an API to create custom operators whose
implementations are backed by triton kernels. While user-defined triton
kernels work out-of-the-box with triton kernels, you may wish to
construct a custom operator if you need to compose with other PyTorch
subsystems, like Tensor subclasses or vmap.

I'm hoping to get design feedback on this and ship it so that we can
begin experimenting with customers.

Test Plan:
- new tests

ghstack-source-id: b652143589360aad10eed0bfc1c87763abda3655
Pull Request resolved: https://github.com/pytorch/pytorch/pull/130637