[inductor] Fix var_to_range in IndexPropagation
The current code assumes that indirect variables will be created by the
same `IndexPropagation` instance, however that isn't true in the case of
masked sub-blocks where we take in variables from the parent block.

This fixes the issue by moving the var range information up to the
`LoopBody` object where it can be shared by all sub-blocks.

ghstack-source-id: f0b5a7f69d2569339e2825400e11f130c48cfdd2
Pull Request resolved: https://github.com/pytorch/pytorch/pull/130984