[inductor] Avoid fallback case for custom scan op lowering
We currently can't generate split scans when there are multiple scan
values, so we normally fall back to ATen. However, for the higher order
scan op, we can't fallback so it makes sense to just generate the slower
kernel anyway. This avoids having special shapes where we fail to
codegen.

ghstack-source-id: 5346730a92c80cde944ad34529bd33180ab945c8
Pull Request resolved: https://github.com/pytorch/pytorch/pull/130936