[HOP] Don't send HOPs to torch_dispatch
I regretted the decision in
https://github.com/pytorch/pytorch/pull/130606. Most user
torch_dispatchs don't have enough to actually handle the HOP correctly,
so for now I'd prefer that users explicitly define the interaction
between the HOP and their torch_dispatch class.

An example is FlopCounterMode: if we allow HOPs to get passed to it, it
will ignore auto_functionalized(mm) by default but it will record flops
for mm, which is weird.

Test Plan:
- tests

ghstack-source-id: dc8ef9f7c3960323a99f6c02079fe4f6e51736df
Pull Request resolved: https://github.com/pytorch/pytorch/pull/131370