Add flag to ignore unsupported triton.autotune args in user-written kernel compilation
Summary: We currently don't support some of the `triton.autotune` arguments when compiling user-written Triton kernels with PT2. In this PR, we're adding a flag to circumvent it. This is to unblock internal compilation in some cases. The flag is supplied with the docs mentioning why it is not a good idea to set it.

Test Plan:

```
python test/inductor/test_triton_kernels.py -k test_triton_kernel_
autotune_with_unsupported_args
...
----------------------------------------------------------------------
Ran 3 tests in 3.636s

OK
```

Reviewers:

Subscribers:

Tasks:

Tags:

ghstack-source-id: 8d328024e86d840b66c8f7c8208ab7dc10739cac
Pull Request resolved: https://github.com/pytorch/pytorch/pull/131431