Add support to choose AMI
This extends the runner determinator to be able to opt-in to keywords
to provide additional options when determining which systems to run
jobs on.

This change creates a generic get_optin_feature() which hopefully will
be useful to handle additional future features that we might want to
experiment with.

This closes #249.

Signed-off-by: Thanh Ha <thanh.ha@linuxfoundation.org>